[runtime/ttnn] Honor host-buffer strides when creating owned tensors (#8697)
### Ticket
fixes: #8622

### Problem description
`createOwnedHostTensor` accepted a `stride` argument but ignored it,
which meant that the host buffer was always read linearly as dense
row-major. When a frontend hands us a non-contiguous host buffer (a
transposed or sliced view), the data was interpreted as contiguous and
silently corrupted (Note: this is currently only possible trough `jax`
path, because we are forcing `pytorch` tensors to be contiguous in
frontend).

### What's changed
`createOwnedTTNNTensor` now detects a non-contiguous layout from the
element strides (`isNonContiguous`) and gathers the strided source into
a dense, row-major contiguous buffer before constructing the tensor
(`gatherContiguousBytes`), then builds the owned tensor as usual.
Contiguous inputs are unchanged.

+++ runtime/lib/distributed/controller/command_factory.cpp
@@ -214,7 +214,7 @@ uint64_t CommandFactory::buildGetMeshShapeCommand(
 uint64_t CommandFactory::buildCreateHostTensorCommand(
     ::flatbuffers::FlatBufferBuilder &fbb,
     const ::tt::runtime::Tensor &outputTensor, const void *data,
-    const std::vector<uint32_t> &shape, const std::vector<uint32_t> &stride,
+    const std::vector<uint32_t> &shape, const std::vector<int64_t> &stride,
     uint32_t itemSize, ::tt::target::DataType dataType) {
 
   LOG_ASSERT(fbb.GetSize() == 0, "Flatbuffer builder must be empty");
@@ -226,7 +226,7 @@ uint64_t CommandFactory::buildCreateHostTensorCommand(
   auto dataVec =
       fbb.CreateVector<uint8_t>(static_cast<const uint8_t *>(data), numBytes);
   auto shapeVec = fbb.CreateVector<uint32_t>(shape.data(), shape.size());
-  auto strideVec = fbb.CreateVector<uint32_t>(stride.data(), stride.size());
+  auto strideVec = fbb.CreateVector<int64_t>(stride.data(), stride.size());
 
   uint64_t commandId =
       BUILD_COMMAND(CreateHostTensor, fbb, outputTensor.getGlobalId(), dataVec,
